+++ tests/core/pm-module/test_registry_integration.py
@@ -25,7 +25,7 @@ from web3.pm import (
 def fresh_w3():
     w3 = Web3(Web3.EthereumTesterProvider())
     w3.eth.default_account = w3.eth.accounts[0]
-    w3.eth.defaultContractFactory = LinkableContract
+    w3.eth._default_contract_factory = LinkableContract
     w3.enable_unstable_package_management_api()
     return w3
