+++ web3/_utils/filters.py
@@ -376,7 +376,7 @@ def select_filter_method(
     if_new_block_filter: RPCEndpoint,
     if_new_pending_transaction_filter: RPCEndpoint,
     if_new_filter: RPCEndpoint,
-) -> RPCEndpoint:
+) -> Optional[RPCEndpoint]:
 
     if is_string(value):
         if value == "latest":
