+++ web3/eth/__init__.py
@@ -1,0 +1,10 @@
+from .async_eth import (  # noqa: F401
+    AsyncEth,
+)
+from .base import (  # noqa: F401
+    BaseEth,
+)
+from .main import (  # noqa: F401
+    Eth,
+    Contract,
+)
